+++ test/registered/kernels/test_nsa_indexer.py
@@ -11,7 +11,7 @@ from sglang.test.ci.ci_register import register_cuda_ci
 _dp_attn.get_attention_tp_size = lambda: 1  # TP size = 1 for unit test
 
 from sglang.srt.configs.model_config import AttentionArch
-from sglang.srt.layers.attention.dsa.nsa_indexer import (
+from sglang.srt.layers.attention.dsa.dsa_indexer import (
     BaseIndexerMetadata,
     Indexer,
     rotate_activation,
@@ -417,7 +417,7 @@ class TestNSAIndexer(CustomTestCase):
             "Output should have padding or exact topk size",
         )
 
-    @patch("sglang.srt.layers.attention.dsa.nsa_indexer.deep_gemm")
+    @patch("sglang.srt.layers.attention.dsa.dsa_indexer.deep_gemm")
     def test_indexer_basic_creation(self, mock_deep_gemm):
         """Test basic indexer creation and initialization."""
         mock_deep_gemm.get_num_sms.return_value = 132
@@ -431,7 +431,7 @@ class TestNSAIndexer(CustomTestCase):
         self.assertEqual(indexer.index_topk, self.config["index_topk"])
         self.assertEqual(indexer.layer_id, self.config["layer_id"])
 
-    @patch("sglang.srt.layers.attention.dsa.nsa_indexer.deep_gemm")
+    @patch("sglang.srt.layers.attention.dsa.dsa_indexer.deep_gemm")
     @patch("sglang.srt.layers.attention.dsa.triton_kernel.act_quant")
     def test_forward_extend_mode(self, mock_act_quant, mock_deep_gemm):
         """Test indexer forward pass in extend mode."""
@@ -513,7 +513,7 @@ class TestNSAIndexer(CustomTestCase):
             topk_indices, self.batch_size, self.seq_len, self.config["index_topk"]
         )
 
-    @patch("sglang.srt.layers.attention.dsa.nsa_indexer.deep_gemm")
+    @patch("sglang.srt.layers.attention.dsa.dsa_indexer.deep_gemm")
     @patch("sglang.srt.layers.attention.dsa.triton_kernel.act_quant")
     def test_forward_decode_mode(self, mock_act_quant, mock_deep_gemm):
         """Test indexer forward pass in decode mode."""
@@ -627,7 +627,7 @@ class TestNSAIndexer(CustomTestCase):
         self.assertEqual(topk_indices.shape, (batch_size, topk))
 
     # TODO: enable this test after indexer accuracy aligned
-    # @patch("sglang.srt.layers.attention.dsa.nsa_indexer.deep_gemm")
+    # @patch("sglang.srt.layers.attention.dsa.dsa_indexer.deep_gemm")
     # def test_indexer_with_different_topk(self, mock_deep_gemm):
     #     """Test indexer with different topk values."""
     #     mock_deep_gemm.get_num_sms.return_value = 132
@@ -637,7 +637,7 @@ class TestNSAIndexer(CustomTestCase):
     #             indexer = self._create_indexer(index_topk=topk)
     #             self.assertEqual(indexer.index_topk, topk)
 
-    @patch("sglang.srt.layers.attention.dsa.nsa_indexer.deep_gemm")
+    @patch("sglang.srt.layers.attention.dsa.dsa_indexer.deep_gemm")
     def test_indexer_with_fused_wk(self, mock_deep_gemm):
         """Test indexer creation with fused wk and weights projection."""
         mock_deep_gemm.get_num_sms.return_value = 132
@@ -647,7 +647,7 @@ class TestNSAIndexer(CustomTestCase):
         indexer = self._create_indexer()
         self.assertIsNotNone(indexer)
 
-    @patch("sglang.srt.layers.attention.dsa.nsa_indexer.deep_gemm")
+    @patch("sglang.srt.layers.attention.dsa.dsa_indexer.deep_gemm")
     def test_indexer_with_alt_stream(self, mock_deep_gemm):
         """Test indexer creation with alternative CUDA stream."""
         mock_deep_gemm.get_num_sms.return_value = 132
